fix: update-hook-socket-get-rate
